old_stuff/requirements.txt
playwright
supabase
python-dotenv
pytz
